redux set up for artist list
