--- src/actions/Form.js
@@ -1,7 +1,0 @@
-import { getForm } from '../services/formApi';
-
-export const FETCH_ARTISTS = 'FETCH_ARTISTS';
-export const fetchArtists = () => ({
-  type: FETCH_ARTISTS,
-  payload: getForm()
-});
