+++ src/reducers/artists.js
@@ -1,30 +1,15 @@
-
+import { FETCH_ARTISTS } from '../actions/artists';
 
 const initialState = {
-  artists: [
-    {
-      name: 'Tyler',
-      url: 'https://proxy.duckduckgo.com/iu/?u=http%3A%2F%2Fi.ytimg.com%2Fvi%2F7pgqf1hQ648%2Fmaxresdefault.jpg&f=1'
-    },
-    {
-      name: 'Ivan',
-      url: 'https://proxy.duckduckgo.com/iu/?u=http%3A%2F%2Fi.ytimg.com%2Fvi%2F7pgqf1hQ648%2Fmaxresdefault.jpg&f=1'
-    },
-    {
-      name: 'Sophie',
-      url: 'https://proxy.duckduckgo.com/iu/?u=http%3A%2F%2Fi.ytimg.com%2Fvi%2F7pgqf1hQ648%2Fmaxresdefault.jpg&f=1'
-    },
-    {
-      name: 'Johnny',
-      url: 'https://proxy.duckduckgo.com/iu/?u=http%3A%2F%2Fi.ytimg.com%2Fvi%2F7pgqf1hQ648%2Fmaxresdefault.jpg&f=1'
-    },
-    {
-      name: 'Michael',
-      url: 'https://proxy.duckduckgo.com/iu/?u=http%3A%2F%2Fi.ytimg.com%2Fvi%2F7pgqf1hQ648%2Fmaxresdefault.jpg&f=1'
-    }
-  ]
+  artists: []
 };
 
-export default function reducer(state = initialState, action) {
-  return state;
+export default function reducer(state = initialState, { type, payload }) {
+  switch(type) {
+    case FETCH_ARTISTS:
+      return { ...state,
+        artists: payload
+      };
+    default: return state;
+  }
 }
