--- src/reducers/form.js
@@ -1,15 +1,0 @@
-import { FETCH_ARTISTS } from '../actions/Form';
-
-const initialState = {
-  artists: []
-};
-
-export default function reducer(state = initialState, { type, payload }) {
-  switch(type) {
-    case FETCH_ARTISTS:
-      return { ...state,
-        artists: payload
-      };
-    default: return state;
-  }
-}
