+++ src/reducers/index.js
@@ -1,9 +1,7 @@
 import { combineReducers } from 'redux';
 import artists from './artists';
-import artistArray from './form';
 
 export default combineReducers({
-  artists,
-  artistArray
+  artists
 
 });
