--- src/services/formApi.js
@@ -1,7 +1,0 @@
-export const getForm = () => {
-  return Promise.resolve([ 
-    { artistName: 'coldplay', email:'coldplay@gmail.com' },
-    { artistName: 'Blueface', email:'yeaAyt@gmail.com' },
-    { artistName: 'Tekashi69', email:'traeway@gmail.com' }
-  ]);
-};
